+++ packages/react/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @dropinnodex/react
 
+## 0.12.1
+
+### Patch Changes
+
+- Updated dependencies [2a5745f]
+  - @dropinnodex/client@0.9.0
+
 ## 0.12.0
 
 ### Minor Changes